requirements.txt
Package             Version 
------------------- --------
contourpy           1.0.7   
cycler              0.11.0  
fonttools           4.39.2  
importlib-resources 5.12.0  
kiwisolver          1.4.4   
matplotlib          3.7.1   
numpy               1.24.2  
opencv-python       4.7.0.72
packaging           23.0    
Pillow              9.4.0   
pip                 20.0.2  
pkg-resources       0.0.0   
pyparsing           3.0.9   
python-dateutil     2.8.2   
python-dotenv       1.0.0   
setuptools          44.0.0  
six                 1.16.0  
zipp                3.15.0  
